backend/requirements.txt
fastapi==0.135.3
uvicorn[standard]==0.42.0
sqlalchemy==2.0.48
pydantic==2.12.5
python-multipart==0.0.22
pypdf==6.9.2
reportlab==4.4.10
python-dotenv==1.2.2
openai==2.30.0
